PS-2653 Remove TableWriterV1, remove factory layer
